# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test work 1",
        "description": "Test work of Rest Api by Laravel 11 platform",
        "contact": {
            "name": "ASt",
            "email": "alexandr.statut@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Dev API server"
        },
        {
            "url": "http://127.0.0.1:8001/api",
            "description": "Prod API server"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Авторизация"
                ],
                "summary": "Api: Авторизация",
                "description": "Получение персонального токена",
                "operationId": "auth.login",
                "requestBody": {
                    "description": "Учетные данные",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "title": "Логин (E-mail)",
                                        "type": "string",
                                        "nullable": "false"
                                    },
                                    "password": {
                                        "title": "Пароль",
                                        "type": "string",
                                        "nullable": "false"
                                    }
                                },
                                "type": "object"
                            },
                            "examples": {
                                "Admin": {
                                    "summary": "Admin",
                                    "value": {
                                        "email": "admin@test.com",
                                        "password": "admin"
                                    }
                                },
                                "Moderator": {
                                    "summary": "Moderator",
                                    "value": {
                                        "email": "employee@test.com",
                                        "password": "employee"
                                    }
                                },
                                "User": {
                                    "summary": "User",
                                    "value": {
                                        "email": "user1@test.com",
                                        "password": "user"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/TokenResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ErrorResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ErrorResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "server not available",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ErrorResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/me": {
            "post": {
                "tags": [
                    "Авторизация"
                ],
                "summary": "Api: Авторизованный пользователь",
                "description": "Информация об авторизованном пользователе",
                "operationId": "auth.me",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/UserResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ErrorResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ErrorResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ErrorResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "server not available",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ErrorResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Авторизация"
                ],
                "summary": "Api: Logout",
                "description": "Logout",
                "operationId": "auth.logout",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/MessageResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ErrorResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ErrorResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ErrorResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "server not available",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ErrorResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorResource": {
                "title": "Ответ с ошибкой",
                "properties": {
                    "status": {
                        "title": "Статус",
                        "type": "boolean",
                        "format": "int64",
                        "example": "false"
                    },
                    "code": {
                        "title": "Код",
                        "type": "integer",
                        "format": "int64",
                        "example": "500"
                    },
                    "msg": {
                        "title": "Сообщение",
                        "type": "string",
                        "example": "Test"
                    }
                },
                "type": "object"
            },
...
            "MessageResource": {
                "title": "Простой ответ",
                "properties": {
                    "status": {
                        "title": "Статус",
                        "description": "Transform the resource into an array.",
                        "type": "boolean",
                        "example": "true"
                    },
                    "msg": {
                        "title": "Сообщение",
                        "type": "string",
                        "example": "Test"
                    }
                },
                "type": "object"
            },
...
            "TokenResource": {
                "title": "Персональный токен",
                "properties": {
                    "token": {
                        "title": "Токен",
                        "description": "Transform the resource into an array.",
                        "type": "string",
                        "example": "1|abcdefjhijklmn"
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "title": "Карточка пользователя",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "Transform the resource into an array.",
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "name": {
                        "title": "ФИО",
                        "type": "string",
                        "example": "Test User"
                    },
                    "email": {
                        "title": "Логин",
                        "type": "string",
                        "example": "test@test.com"
                    },
                    "status": {
                        "title": "Статус",
                        "type": "integer",
                        "example": "1"
                    },
                    "created": {
                        "title": "Создана",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-12-01 12:00:00"
                    },
                    "updated": {
                        "title": "Обновлена",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-12-01 12:00:00"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Your token providing after auth in Api",
                "name": "Api Client",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}